+++ packages/angular-cli/commands/build.ts
@@ -11,6 +11,10 @@ export interface BuildOptions {
   supressSizes: boolean;
   baseHref?: string;
   aot?: boolean;
+  sourcemap?: boolean;
+  vendorChunk?: boolean;
+  verbose?: boolean;
+  progress?: boolean;
 }
 
 const BuildCommand = Command.extend({
@@ -31,7 +35,11 @@ const BuildCommand = Command.extend({
     { name: 'watcher',        type: String },
     { name: 'suppress-sizes', type: Boolean, default: false },
     { name: 'base-href',      type: String,  default: null, aliases: ['bh'] },
-    { name: 'aot',            type: Boolean, default: false }
+    { name: 'aot',            type: Boolean, default: false },
+    { name: 'sourcemap',      type: Boolean, default: true, aliases: ['sm'] },
+    { name: 'vendor-chunk',   type: Boolean, default: true },
+    { name: 'verbose',        type: Boolean, default: false },
+    { name: 'progress',       type: Boolean, default: true }
   ],
 
   run: function (commandOptions: BuildOptions) {