+++ packages/angular-cli/models/webpack-build-utils.ts
@@ -4,20 +4,30 @@ export const ngAppResolve = (resolvePath: string): string => {
   return path.resolve(process.cwd(), resolvePath);
 };
 
-export const webpackOutputOptions = {
+const webpackOutputOptions = {
   colors: true,
-  chunks: true,
-  modules: false,
-  reasons: false,
-  chunkModules: false
-};
-
-export const webpackDevServerOutputOptions = {
-  assets: true,
-  colors: true,
-  version: true,
   hash: true,
   timings: true,
-  chunks: false,
-  chunkModules: false
+  chunks: true,
+  chunkModules: false,
+  children: false, // listing all children is very noisy in AOT and hides warnings/errors
+  modules: false,
+  reasons: false,
+  warnings: true,
+  assets: false, // listing all assets is very noisy when using assets directories
+  version: false
 };
+
+const verboseWebpackOutputOptions = {
+  children: true,
+  assets: true,
+  version: true,
+  reasons: true,
+  chunkModules: false // TODO: set to true when console to file output is fixed
+};
+
+export function getWebpackStatsConfig(verbose = false) {
+    return verbose
+           ? Object.assign(webpackOutputOptions, verboseWebpackOutputOptions)
+           : webpackOutputOptions;
+}