+++ packages/angular-cli/plugins/karma.js
@@ -12,7 +12,9 @@ const init = (config) => {
   const environment = config.angularCli.environment || 'dev';
   const testConfig = {
     codeCoverage: config.angularCli.codeCoverage || false,
-    lint: config.angularCli.lint || false
+    lint: config.angularCli.lint || false,
+    sourcemap: config.angularCli.sourcemap,
+    progress: config.angularCli.progress
   }
 
   // add webpack config