+++ packages/angular-cli/tasks/build-webpack.ts
@@ -4,8 +4,9 @@ const Task = require('../ember-cli/lib/models/task');
 import * as webpack from 'webpack';
 import { BuildOptions } from '../commands/build';
 import { NgCliWebpackConfig } from '../models/webpack-config';
-import { webpackOutputOptions } from '../models/';
+import { getWebpackStatsConfig } from '../models/';
 import { CliConfig } from '../models/config';
+
 
 // Configure build and output;
 let lastHash: any = null;
@@ -23,20 +24,22 @@ export default <any>Task.extend({
       runTaskOptions.environment,
       outputDir,
       runTaskOptions.baseHref,
-      runTaskOptions.aot
+      runTaskOptions.aot,
+      runTaskOptions.sourcemap,
+      runTaskOptions.vendorChunk,
+      runTaskOptions.verbose,
+      runTaskOptions.progress
     ).config;
 
     const webpackCompiler: any = webpack(config);
 
-    const ProgressPlugin  = require('webpack/lib/ProgressPlugin');
-
-    webpackCompiler.apply(new ProgressPlugin({
-      profile: true
-    }));
+    const statsConfig = getWebpackStatsConfig(runTaskOptions.verbose);
 
     return new Promise((resolve, reject) => {
       webpackCompiler.run((err: any, stats: any) => {
-        if (err) { return reject(err); }
+        if (err) {
+          return reject(err);
+        }
 
         // Don't keep cache
         // TODO: Make conditional if using --watch
@@ -44,11 +47,21 @@ export default <any>Task.extend({
 
         if (stats.hash !== lastHash) {
           lastHash = stats.hash;
-          process.stdout.write(stats.toString(webpackOutputOptions) + '\n');
+          process.stdout.write(stats.toString(statsConfig) + '\n');
         }
 
-        return stats.hasErrors() ? reject() : resolve();
+        if (stats.hasErrors()) {
+          reject();
+        } else {
+          resolve();
+        }
       });
+    })
+    .catch((err: Error) => {
+      if (err) {
+        this.ui.writeError('\nAn error occured during the build:\n' + ((err && err.stack) || err));
+      }
+      throw err;
     });
   }
 });