+++ tests/e2e/tests/third-party/bootstrap.ts
@@ -25,6 +25,7 @@ export default function() {
       <script type="text/javascript" src="inline.bundle.js"></script>
       <script type="text/javascript" src="styles.bundle.js"></script>
       <script type="text/javascript" src="scripts.bundle.js"></script>
+      <script type="text/javascript" src="vendor.bundle.js"></script>
       <script type="text/javascript" src="main.bundle.js"></script>
     `));
 }